Portefølje attempt

+++ app/src/main/java/com/example/kind1/data/Donation.kt
@@ -1,7 +1,7 @@
 package com.example.kind1.data
 
 data class Donation(
-    var amount: Int = 0,
+    var amount: Long = 0,
     val orgName: String = "",
     var username: String = "",
     val isMonthly: Boolean = true,
